requirements.txt
# Shared deps for Trip Planner + Country Financial Insights agents
streamlit>=1.28.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
python-dotenv>=1.0.0
requests>=2.31.0
yfinance>=0.2.40

# Optional: for real flight/hotel data in trip_planner
amadeus>=2.0.0
